+++ services/src/pro/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod openapi_examples;
 // #[cfg(test)] /// TODO: currently also used in quota_check bench. Maybe just copy the code there.
 pub mod tests;
