+++ services/src/util/mod.rs
@@ -10,6 +10,7 @@ pub mod apidoc;
 pub mod config;
 pub mod extractors;
 pub mod identifiers;
+pub mod openapi_examples;
 pub mod operators;
 pub mod parsing;
 pub mod postgres;
